написал энд поинты для получения количества участников в группе

+++ src/main/java/com/example/prodBack8/services/TaskService.java
@@ -11,5 +11,5 @@ public interface TaskService {
     Integer getDistributionGroupByUserId(UserEntity entity);
     Integer getMaxSessionDurationGroupByUserId(UserEntity entity);
     String getAllowedTimeGroupByUserId(UserEntity entity);
-    Long getCountMembersGroupById(UserEntity entity);
+    Integer getCountMembersGroupById(UserEntity entity);
 }
